etc: Add reason_code, video_id for report

+++ src/main/java/com/jocoos/mybeautip/video/Video.java
@@ -86,7 +86,7 @@ public class Video {
   private Integer orderCount;
   
   @Column
-  private Integer reportCount;
+  private Long reportCount;
 
   @Column(nullable = false)
   private Integer relatedGoodsCount;
